move modals to tera-project-menu
